pyproject.toml
[project]
name = "agentic-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "googlemaps>=4.10.0",
    "langchain>=0.3.26",
    "langchain-cerebras>=0.5.0",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "typing>=3.10.0.0",
    "uvloop>=0.21.0",
]
